Add support for 'emails' key in ServiceChannel response
Basically replicates commit 8f37b047bb95224f67d31fabf9144db0b90cb411 in the new
codebase.

This adds support for the 'emails' key that exists in
ServiceChannel/<uuid> as well as ServiceChannel/list responses. For our
purposes it's enough to support this in the ServiceLocationChannel type
only, at least for now.