Add phpexperts/docker-phan and config for phan
These additions help to ensure validity when running the target php version
which is 8.1. The minimum severity has been set to critical since the code
generated by the openapi-generator still results in a lot of warnings on lower
levels.

The docker-phan package was added instead of the phan/phan package to get rid
of the requirement to install ext-ast in the PHP system.